+++ src/main/java/rizzler/parser/DateTimeParser.java
@@ -1,4 +1,4 @@
-package rizzler.ui.parser;
+package rizzler.parser;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
